Cart init

+++ Ecommerce/src/main/java/com/bourgedetrembleur/hepl/repository/ArticleRepository.java
@@ -3,7 +3,7 @@ package com.bourgedetrembleur.hepl.repository;
 import com.bourgedetrembleur.hepl.model.Article;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ArticleRepository extends PagingAndSortingRepository<Article, Long>
+public interface ArticleRepository extends PagingAndSortingRepository<Article, Integer>
 {
     public static final int PAGE_SIZE = 5;
 }
